Update package.json to allow devs to complete postinstall script. Split components out into dedicated files. Convert components that were able into pure renders.
