+++ app/scripts/components/Sidebar.jsx
@@ -1,0 +1,17 @@
+import React from 'react'
+
+let Sidebar = (props) => {
+  render() {
+    return(
+        <div id="sidebar-wrapper">
+          <ul className="sidebar-nav">
+            <li className="sidebar-brand"><a href="#"></a></li>
+            <li><a target='_blank' href="https://github.com/alanbsmith/react-retro-sockets">View the Repo</a></li>
+            <li><a target='_blank' href="https://twitter.com/_alanbsmith">On Twitter</a></li>
+          </ul>
+        </div>
+    )
+  }
+};
+
+export default Sidebar
