+++ package.json
@@ -24,5 +24,8 @@
     "socket.io": "^1.3.7",
     "style-loader": "^0.13.0"
   },
-  "devDependencies": {}
+  "devDependencies": {
+    "react-dom": "^0.14.7",
+    "webpack": "^1.12.14"
+  }
 }
